requirements.txt
Django
wagtail
djangorestframework 
djangorestframework-simplejwt
django-cors-headers
mysqlclient  